tree simplified
