--- screens/main/chat/components/common/Loader/styles.js
@@ -1,16 +1,0 @@
-import { StyleSheet } from 'react-native'
-
-export default StyleSheet.create({
-  container: {
-    position: 'absolute',
-
-    top: 0,
-    left: 0,
-
-    height: '100%',
-    width: '100%',
-
-    justifyContent: 'center',
-    alignItems: 'center'
-  }
-})
